+++ packages/roosterjs-editor-api/lib/table/insertTable.ts
@@ -1,5 +1,25 @@
-import { ChangeSource, IEditor, PositionType, TableFormat } from 'roosterjs-editor-types';
-import { Position, VTable } from 'roosterjs-editor-dom';
+import { Position, saveTableInfo, VTable } from 'roosterjs-editor-dom';
+import {
+    ChangeSource,
+    IEditor,
+    PositionType,
+    TableBorderFormat,
+    TableFormat,
+} from 'roosterjs-editor-types';
+
+const DEFAULT_FORMAT: TableFormat = {
+    topBorderColor: '#ABABAB',
+    bottomBorderColor: '#ABABAB',
+    verticalBorderColor: '#ABABAB',
+    hasHeaderRow: false,
+    hasFirstColumn: false,
+    hasBandedRows: false,
+    hasBandedColumns: false,
+    bgColorEven: null,
+    bgColorOdd: '#ABABAB20',
+    headerRowColor: '#ABABAB',
+    tableBorderFormat: TableBorderFormat.DEFAULT,
+};
 
 /**
  * Insert table into editor at current selection
@@ -35,8 +55,9 @@ export default function insertTable(
 
     editor.focus();
     editor.addUndoSnapshot(() => {
+        saveTableInfo(table, format || DEFAULT_FORMAT);
         let vtable = new VTable(table);
-        vtable.applyFormat(format);
+        vtable.applyFormat(format || DEFAULT_FORMAT);
         vtable.writeBack();
         editor.insertNode(fragment);
         editor.runAsync(editor =>
