+++ packages/roosterjs-editor-dom/test/utils/applyTableFormatTest.ts
@@ -1,0 +1,36 @@
+import applyTableFormat from '../../lib/utils/applyTableFormat';
+import VTable from '../../lib/table/VTable';
+import { itChromeOnly } from '../DomTestHelper';
+import { TableFormat } from 'roosterjs-editor-types';
+
+const format: TableFormat = {
+    topBorderColor: '#0C64C0',
+    bottomBorderColor: '#0C64C0',
+    verticalBorderColor: '#0C64C0',
+    bgColorEven: '#0C64C020',
+    bgColorOdd: null,
+    headerRowColor: null,
+    tableBorderFormat: 0,
+    hasHeaderRow: false,
+    hasFirstColumn: false,
+    hasBandedRows: false,
+    hasBandedColumns: false,
+};
+
+describe('applyTableFormat', () => {
+    let table =
+        '<table cellspacing="0" cellpadding="1" id="id1" data-rooster-table-info="{&quot;topBorderColor&quot;:&quot;#ABABAB&quot;,&quot;bottomBorderColor&quot;:&quot;#ABABAB&quot;,&quot;verticalBorderColor&quot;:&quot;#ABABAB&quot;,&quot;hasHeaderRow&quot;:false,&quot;hasFirstColumn&quot;:false,&quot;hasBandedRows&quot;:false,&quot;hasBandedColumns&quot;:false,&quot;bgColorEven&quot;:null,&quot;bgColorOdd&quot;:&quot;#ABABAB20&quot;,&quot;headerRowColor&quot;:&quot;#ABABAB&quot;,&quot;tableBorderFormat&quot;:0}" style="border-collapse: collapse;"><tbody><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(171, 171, 171);"><br></td></tr></tbody></table>';
+    let expectedTableChrome =
+        '<table cellspacing="0" cellpadding="1" id="id1" data-rooster-table-info="{&quot;topBorderColor&quot;:&quot;#0C64C0&quot;,&quot;bottomBorderColor&quot;:&quot;#0C64C0&quot;,&quot;verticalBorderColor&quot;:&quot;#0C64C0&quot;,&quot;hasBandedRows&quot;:false,&quot;bgColorEven&quot;:null,&quot;bgColorOdd&quot;:&quot;#0C64C020&quot;,&quot;hasBandedColumns&quot;:false,&quot;hasHeaderRow&quot;:false,&quot;headerRowColor&quot;:&quot;#0C64C0&quot;,&quot;hasFirstColumn&quot;:false,&quot;tableBorderFormat&quot;:0}" style="border-collapse: collapse;"><tbody><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td></tr><tr><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);" scope=""><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td><td style="width: 100px; border-width: 1px; border-style: solid; border-color: rgb(12, 100, 192);"><br></td></tr></tbody></table>';
+    let div = document.createElement('div');
+    document.body.appendChild(div);
+    const id = 'id1';
+    div.innerHTML = table;
+    let node = document.getElementById(id) as HTMLTableElement;
+    let vTable = new VTable(node);
+    applyTableFormat(vTable.cells, format);
+    itChromeOnly('should return a styled table CHROME', () => {
+        expect(div.innerHTML).toBe(expectedTableChrome);
+    });
+    document.body.removeChild(div);
+});
